requirements.txt
streamlit
google.generativeai
pandas
plotly
requests